fix: unify and correct redux store
