--- src/app/StoreProvider.tsx
@@ -1,19 +1,0 @@
-'use client';
-
-import { useRef } from 'react';
-import { Provider } from 'react-redux';
-import { makeStore, AppStore } from '@/lib/store';
-
-export default function StoreProvider({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const storeRef = useRef<AppStore | null>(null);
-
-  if (!storeRef.current) {
-    storeRef.current = makeStore();
-  }
-
-  return <Provider store={storeRef.current}>{children}</Provider>;
-}
