+++ src/components/StoreProvider/StoreProvider.tsx
@@ -1,8 +1,19 @@
 'use client';
 
-import { store } from '@/store/store';
+import { AppStore, makeStore } from '@/store/store';
+import { useRef } from 'react';
 import { Provider } from 'react-redux';
 
-export function StoreProvider({ children }: { children: React.ReactNode }) {
-  return <Provider store={store}>{children}</Provider>;
+export default function StoreProvider({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const storeRef = useRef<AppStore | null>(null);
+
+  if (!storeRef.current) {
+    storeRef.current = makeStore();
+  }
+
+  return <Provider store={storeRef.current}>{children}</Provider>;
 }
